DRAMA MU BOT/requirements.txt
fastapi
uvicorn
psycopg2-binary
midtransclient
gunicorn